# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Web Appliance Dashboard API
  description: RESTful API für die Verwaltung von Web-Appliances, Services und SSH-Verbindungen
  version: 1.0.3
  contact:
    name: Alf Lewerken
    email: support@example.com
    url: https://github.com/alflewerken/web-appliance-dashboard
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.appliance-dashboard.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
...
security:
  - bearerAuth: []
paths:
...
  /terminal/sessions:
    get:
      tags:
        - Terminal
      summary: Terminal-Sitzungen auflisten
      responses:
        '200':
          description: Sitzungsliste
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sessions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        appliance_id:
                          type: integer
                        appliance_name:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        last_activity:
                          type: string
                          format: date-time

    post:
      tags:
        - Terminal
      summary: Terminal-Sitzung erstellen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - appliance_id
              properties:
                appliance_id:
                  type: integer
                cols:
                  type: integer
                  default: 80
                rows:
                  type: integer
                  default: 24
      responses:
        '201':
          description: Sitzung erstellt
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  session:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      websocket_url:
                        type: string
                      created_at:
                        type: string
                        format: date-time
  /terminal/sessions/{id}:
    delete:
      tags:
        - Terminal
      summary: Terminal-Sitzung beenden
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sitzung beendet
...